ToConcentration implementation (#10)
This implements a ToConcentration action, which makes the concentration of (non-mix) components in a mix equal to some value. It is distinct from FixedConcentration in that the latter adds a certain concentration, regardless of what might be in the mix from elsewhere.

This action is useful if, eg, you have some mixes with MgCl₂ or polysorbate 20 added, and other without them, and you want to have a mix with those at specific concentrations, taking into account what has been added from the various components.